chore: more refactor work
