+++ auth_api/common.py
@@ -4,12 +4,32 @@ import re
 import requests
 import sys
 import yaml
+import datetime
+import sshpubkeys
+import jinja2
+import smtplib
+import ssl
+import os
+import traceback
+
+from flask import current_app
+from flaskext.mysql import MySQL
 
 API_VERSION = 1
 
-def get_config():
+def get_db():
+    mysql = MySQL()
+    mysql.init_app(current_app)
+    db = mysql.get_db()
+    if not db:
+        db = mysql.connect()
+    if not db:
+        raise Exception("Could not connect to database")
+    return db
+
+def get_config(filepath="/etc/auth_api.yaml"):
     try:
-        with open("/etc/auth_api.yaml") as fh:
+        with open(filepath) as fh:
             config = yaml.safe_load(fh)
     except Exception as e:
         sys.stderr.write(f"Failed loading config: {e}\n")
@@ -21,8 +41,21 @@ def validate_user(user):
         return user
     raise argparse.ArgumentTypeError("Invalid user")
 
+
+'''
+Validate SSH public key
+'''
+def validate_ssh_key(type, pub_key, name):
+    ssh_key = sshpubkeys.SSHKey(f"{type} {pub_key} {name}")
+    try:
+        ssh_key.parse()
+    except Exception:
+        return False
+    return True
+
 def api_get(uri):
-    config = get_config()
+    config = current_app.config['authapi']
+
     url = f"https://{config['host']}/v{API_VERSION}/{uri}"
     response = {}
     try:
@@ -36,3 +69,82 @@ def api_get(uri):
         sys.stderr.write(f"Failed fetching {uri}: {e}\n")
         sys.exit(1)
     return response
+
+'''
+Make list of dicts serializable - convert datetimes to unix timestamps
+'''
+def make_serializable(data):
+    output = []
+    for datum in data:
+        my_output = {}
+        for k, v in datum.items():
+            if isinstance(v, datetime.datetime):
+                my_output[k] = int(v.timestamp())
+            else:
+                my_output[k] = v
+        output.append(my_output)
+    return output
+
+
+'''
+Send email notification to user notifying of key changes
+'''
+def send_email(username, service):
+    db = get_db()
+    smtp_config = current_app.config["smtp"]
+
+    try:
+        template_path = os.path.abspath(os.path.join(os.path.dirname(__file__), "../templates"))
+        with open(os.path.join(template_path, "mail_template.j2")) as fh:
+            mail_template = jinja2.Template(fh.read())
+    except Exception:
+        sys.stderr.write("Failed loading mail template:\n")
+        sys.stderr.write(traceback.format_exc())
+        return False
+
+    try:
+        with db.cursor(dictionary=True) as cursor:
+            cursor.execute("SELECT display_name, email FROM users WHERE username = %s", (username,))
+            result = cursor.fetchall()
+    except Exception:
+        sys.stderr.write("Failed retrieving user details from database:\n")
+        sys.stderr.write(traceback.format_exc())
+        return False
+
+    if len(result) < 1:
+        sys.stderr.write(f"Did not find user {username} in database\n")
+        return False
+
+    if result[0]["display_name"] == "":
+        result[0]["display_name"] = username
+
+    if result[0]["email"] == "":
+        sys.stderr.write(f"Did not find email address for {username} in database\n")
+        return False
+
+    try:
+        mail_message = mail_template.render(
+            from_name=smtp_config["from_name"],
+            from_addr=smtp_config["from_addr"],
+            display_name=result[0]["display_name"],
+            user_email=result[0]["email"],
+            service_name=service.upper(),
+            username=username
+        )
+    except Exception:
+        sys.stderr.write("Failed rendering mail message:\n")
+        sys.stderr.write(traceback.format_exc())
+        return False
+
+    try:
+        smtp = smtplib.SMTP(smtp_config["server"], smtp_config["port"])
+        smtp.starttls(context=ssl.create_default_context())
+        smtp.login(smtp_config["username"], smtp_config["password"])
+        smtp.sendmail(smtp_config["from_addr"], result[0]["email"], mail_message)
+        smtp.quit()
+    except Exception:
+        sys.stderr.write("Failed sending mail message:\n")
+        sys.stderr.write(traceback.format_exc())
+        return False
+
+    return True
